requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
asgiref==3.8.1
attrs==24.3.0
BareNecessities==0.2.8
bcrypt==4.2.1
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==41.0.3
DateTime==5.5
Django==3.1.12
djongo==1.3.7
dnspython==2.3.0
fernet==1.0.1
Flask==2.3.2
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-PyMongo==2.3.0
frozenlist==1.5.0
gunicorn==23.0.0
hcaptcha==0.2.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
Mail==2.1.0
MarkupSafe==3.0.2
mongo==0.2.0
multidict==6.1.0
numpy==2.2.1
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
pillow==11.1.0
propcache==0.2.1
pyaes==1.6.1
pycares==4.5.0
pycparser==2.22
PyJWT==2.10.1
pymongo==3.11.4
PyQRCode==1.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
qrcode==8.0
requests==2.32.3
setuptools==75.8.0
six==1.17.0
sqlparse==0.2.4
tzdata==2024.2
urllib3==2.3.0
Werkzeug==2.3.6
XlsxWriter==3.2.2
yarl==1.18.3
zope.interface==7.2
